+++ app/controllers/projects_controller.rb
@@ -211,7 +211,7 @@ class ProjectsController < ApplicationController
     def project_params
       params.require(:project).permit(
         :name, :description, :acrostic_enabled, :subject_code_name,
-        :show_contacts, :show_documents, :show_posts,
+        :show_contacts, :show_documents, :show_posts, :disable_all_emails,
         # Uploaded Logo
         :logo, :logo_uploaded_at, :logo_cache,
         # Will automatically generate a site if the project has no site