+++ db/migrate/20140130192112_add_disable_all_emails_to_projects.rb
@@ -1,0 +1,5 @@
+class AddDisableAllEmailsToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :disable_all_emails, :boolean, default: false
+  end
+end